Add files via upload
The data set used is  Breast Cancer Wisconsin Diagnostic Database. The dataset includes various information about breast cancer tumors, as well as classification labels of malignant or benign. The dataset has 5 6 9 instances, or data, on 569 tumors and inclues information on 30 attributes, or features, such as the radius of the tumor, texture, smoothness, and area.